+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Datadog AppSec for PHP Release
 
+### v0.2.0
+- The list of response headers transmitted in the trace now contains only:
+	- Content-type
+	- Content-length
+	- Content-encoding
+	- Content-language
+
 ### v0.1.0
 - Initial release